Updated feature with comments
