+++ .classpath
@@ -15,7 +15,7 @@
 	<classpathentry kind="src" path="build"/>
 	<classpathentry kind="src" path="broker"/>
 	<classpathentry kind="src" path="module"/>
-	<classpathentry exported="true" kind="con" path="org.eclipse.jdt.launching.JRE_CONTAINER"/>
+	<classpathentry exported="true" kind="con" path="org.eclipse.jdt.launching.JRE_CONTAINER/org.eclipse.jdt.internal.debug.ui.launcher.StandardVMType/JavaSE-1.8"/>
 	<classpathentry kind="lib" path="main/webapp/WEB-INF/lib/ant-tools-1.8.0.jar"/>
 	<classpathentry kind="lib" path="main/webapp/WEB-INF/lib/arithcode-1.1.jar"/>
 	<classpathentry kind="lib" path="main/webapp/WEB-INF/lib/butterfly-1.0.1.jar"/>
